+++ apps/terminal.js
@@ -12,8 +12,10 @@ function terminal() {
   });
   const protocol = window.location.protocol === "https:" ? "wss" : "ws";
   const host = window.location.hostname;
-  const mrcredentials = JSON.parse(localStorage.getItem('credentials'));
-  const ws = new WebSocket(`${protocol}://${host}:9091/api/terminal?username=${mrcredentials.username}&password=${mrcredentials.password}`);
+  const mrcredentials = JSON.parse(localStorage.getItem("credentials"));
+  const ws = new WebSocket(
+    `${protocol}://${host}:9091/api/terminal?username=${mrcredentials.username}&password=${mrcredentials.password}`,
+  );
   let resizeObserver;
 
   const calculateDimensions = () => {